requirements.txt
streamlit
tensorflow
pandas
numpy
openai
python-dotenv
gspread
oauth2client
openpyxl
geopy
meteostat
